feat: added update task handler

+++ internal/ports/services/tasks/tasks_service.go
@@ -8,9 +8,9 @@ import (
 
 type TaskService interface {
 	CreateTaskService(ctx context.Context, request request.TaskRequest, authorizedId int) (dto.TasksDTO, error)
-	UpdateTaskService(ctx context.Context, id uint8, request request.TaskRequest) (dto.TasksDTO, error)
-	FindTaskByIdService(ctx context.Context, id uint8) (dto.TasksDTO, error)
+	UpdateTaskService(ctx context.Context, id int, request request.TaskRequest) (dto.TasksDTO, error)
+	FindTaskByIdService(ctx context.Context, id int) (dto.TasksDTO, error)
 	FindTaskAllService(ctx context.Context) ([]dto.TasksDTO, error)
-	DeleteTaskService(ctx context.Context, id uint8) error
-	UpdateTaskStatusService(ctx context.Context, id uint8, boolean bool) (dto.TasksDTO, error)
+	DeleteTaskService(ctx context.Context, id int) error
+	UpdateTaskStatusService(ctx context.Context, id int, boolean bool) (dto.TasksDTO, error)
 }
